# tests/openapi/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "examples": [
          {
            "error": {
              "code": "IDEMPOTENCY_KEY_IN_PROGRESS",
              "details": {},
              "http_status": 409,
              "message": "Another request with the same Idempotency-Key is in progress.",
              "trace_id": "req-456"
            }
          }
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorObject"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "ErrorObject": {
        "additionalProperties": false,
        "examples": [
          {
            "code": "IDEMPOTENCY_KEY_CONFLICT",
            "details": {},
            "http_status": 409,
            "message": "Idempotency-Key reused with a different request payload.",
            "trace_id": "req-123"
          }
        ],
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "http_status": {
            "title": "Http Status",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "code",
          "http_status",
          "message"
        ],
        "title": "ErrorObject",
        "type": "object"
      },
...
      "PaginatedEnvelope": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {},
            "title": "Items",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "PaginatedEnvelope",
        "type": "object"
      },
      "QuoteItem": {
        "additionalProperties": false,
        "properties": {
          "as_of": {
            "examples": [
              "2025-10-28T12:34:56Z"
            ],
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "currency": {
            "examples": [
              "USD"
            ],
            "title": "Currency",
            "type": "string"
          },
          "price": {
            "examples": [
              "428.17"
            ],
            "title": "Price",
            "type": "string"
          },
          "ticker": {
            "examples": [
              "MSFT"
            ],
            "title": "Ticker",
            "type": "string"
          },
          "volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              14230000
            ],
            "title": "Volume"
          }
        },
        "required": [
          "as_of",
          "currency",
          "price",
          "ticker"
        ],
        "title": "QuoteItem",
        "type": "object"
      },
      "QuotesBatch": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuoteItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "QuotesBatch",
        "type": "object"
      },
...
      "SuccessEnvelope_QuotesBatch_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QuotesBatch"
          }
        },
        "required": [
          "data"
        ],
        "title": "SuccessEnvelope",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Secure, governed financial data API.",
    "title": "Arche API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/quotes": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "tickers",
            "required": true,
            "schema": {
              "title": "Tickers",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope_QuotesBatch_"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEnvelope"
                    }
                  ],
                  "title": "Response Get  V2 Quotes"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "tags": [
          "Market Data",
          "Market Data"
        ]
      }
    },
    "/v2/quotes/historical": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "from_",
            "required": true,
            "schema": {
              "title": "From ",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "interval",
            "required": true,
            "schema": {
              "enum": [
                "1d",
                "1m"
              ],
              "title": "Interval",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "tickers",
            "required": true,
            "schema": {
              "items": {
                "type": "string"
              },
              "maxItems": 50,
              "minItems": 1,
              "title": "Tickers",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "title": "To",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedEnvelope"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "tags": [
          "Market Data",
          "Market Data"
        ]
      }
    }
  }
}